+++ packages/hono-takibi/src/generators/hono/openapi/component/oneof/generate-oneof-code.test.ts
@@ -1,8 +1,8 @@
 import { describe, expect, it } from 'vitest'
 import { generateOneOfCode } from './generate-oneof-code'
-import { DEFAULT_CONFIG } from '../../../../config'
-import type { Schema } from '../../../../types'
-import type { Config } from '../../../../config'
+import { DEFAULT_CONFIG } from '../../../../../config'
+import type { Schema } from '../../../../../types'
+import type { Config } from '../../../../../config'
 const generateAnyOfCodeTestCases: {
   schema: Schema
   config: Config
