+++ packages/hono-takibi/src/generators/zod/schema/generate-zod-schema.ts
@@ -7,7 +7,7 @@ import { generateZodNumberSchema } from '../generate-zod-number-schema'
 import { generateZodIntegerSchema } from '../generate-zod-integer-schema'
 import { generateAllOfCode } from '../../hono/openapi/component/allof/generate-allof-code'
 import { generateAnyOfCode } from '../../hono/openapi/component/anyof/generate-anyof-code'
-import { generateOneOfCode } from '../../openapi/components/oneof/generate-oneof-code'
+import { generateOneOfCode } from '../../hono/openapi/component/oneof/generate-oneof-code'
 import { getVariableSchemaNameHelper } from '../../../core/helper/get-variable-schema-name-helper'
 import { generateZodObject } from '../generate-zod-object'
 import { generateZodEnum } from '../generate-zod-enum'
